requirements.txt
flask
google-generativeai
better-profanity